Fix the delay for the completion menu to show up
This is caused by the min_keyword_length option in the blink.cmp.
In blink.cmp, it means when to show up the completion menu.
Not when to get the completions, so we add check in
get_completions function to make sure the do not return too many
completions.

This should close #2 .